+++ client/public/index.html
@@ -6,7 +6,6 @@
   <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
   <link rel="manifest" href="%PUBLIC_URL%/manifest.json">
   <link rel="shortcut icon" href="%PUBLIC_URL%/SH.PNG">
-  <link rel="stylesheet" href="%PUBLIC_URL%/font-awesome.min.css">
   <link rel="stylesheet" href="%PUBLIC_URL%/bulma.min.css">
   <script defer src="https://use.fontawesome.com/releases/v5.0.0/js/all.js"></script>
   <title>Smart Hyre</title>
