+++ client/src/pages/companyDashboard/index.css
@@ -1,0 +1,3 @@
+.blackiconcolor {
+  color : #4a4a4a;
+}
